Fix List to work with the new Repetition
This means having to handle the "content" and "contentController" bindings by hand.

The problem is that since now setting the "content" on the Repetition automatically creates a contentController it's not possible to bind to those two properties at the same time even if we only use one.
When the contentController binding is activated "undefined" (because we don't have one in the List if we're using content) will be assigned to it and the content controller that was created for the "content" is lost.