Option element conversion to contentController objects now honor textPropertyPath and valuePropertyPaths, if available
select.reel honors the valuePropertyPath and textPropertyPath when converting a passed in contentController to <option> tags, but does not honor the reverse. If an <option> tag is present, it doesn't convert value and text values into corresponding valuePropertyPath and textPropertyPath properties on the generated objects.

This fixes that by using the valuePropertyPath ad textPropertyPath values when present and defaulting to "value" and "text" when they're not present.

Signed-off-by: Tom Ortega <tom.ortega@gmail.com>