optgroup nodes no longer ignored when reading in from markup
select.reel ignores optgroup when creating objects to stash in the contentController it creates to stash the values; This results in the groups disappearing in the drop down.

This fixes that by creating introducing a groupPropertyPath property and maintaining an offset value to use when calculating selected indexes of selected options.

Signed-off-by: Tom Ortega <tom.ortega@gmail.com>