Commit selection changes only if pointer claimed
This prevents reacting to selection-like events that are actually
intended for child components that have claimed the relevant pointer.

e.g. Pressing a button inside a repetition no longer cane misinterpreted
as making a selection.

If a control should toggle the inclusion of the represented object in
the selection, the control should continue to claim the pointer and 
explicitly bind a value to the selected property.