Change expandParameters to return a richer result
Instead of just returning the labels collision table it returns an object with the labels and element ids that were inserted from the arguments template.
The return is now {labels, labelsCollisions, elementIds, elementIdsCollisions}.

Changed the repetition to use the new result and setup the metadata per object along with the external objects for each template arguments when expanding the inner template into the iteration template.