Add Target class to participate in events
ActiveTarget indicates the logical target the user is focused on much
as the document.activeElement indicates the element the user is
focused on.

activeElement may differ from activeTarget as the activeElement may
have no associated component.

All components are Targets.

This hopefully clarifies some of the half-baked attempts at articulating
this need as evidenced by the component's "acceptsDirectFocus",
and the "parentProperty" to facilitate pointing to the next target
candidate, in addition to a few one-off implementations manually
doing the work of finding the activeComponent.

Specifically, anywhere where we dispatched an event on a component the
user was focused on instead of where the event happened will now be
easier.

The eventManager also provides two new APIs that leverage this new
information, `dispatchFocusedEvent` and `dispatchFocusedEventNamed`
that match their less focused brethren in purpose while using the
activeTarget as the proximal target of the event.

Targets also provide a `nextTarget` property which is used exactly as
the `parentProperty` was to direct distribution of events through a tree
of target candidates.

The implementation thus far reuses the activationEvents concept which
the EventManager already manages. An alternative implementation
could derive the activeTarget from the activeElement and even
drive the prepareForActivationEvent system through focus alone.

That remains a possibility whereas this implementation minimizes
churn by using as much of the existing machinery as possible.