Move eventDispatching off Object.prototype
I didn't even know they were there, now they're on Target. Anybody that
ever plans on being able to be the proximal target or a target during
event distribution will need to be a Target now.

I've found a few places in my limited testing where this was necessary
to introduce in a few object's prototype chain.

The specs are a daunting work in progress.