Move eventListener API off of Object.prototype
More victories in consolidating dispatching responsibilities to
the Target; now the listener registration can be centralized here seeing
as these are the only objects that should be dispatching.

This does elevate Target to be a more vital concept, but it's sensible.

This may prove more disruptive than relocating the dispatching alone,
I haven't rigorously tested.