Change Substitution API and adds spec
The Substitution no longer supports components and is only driven by elements, just like the repetition.

switchComponents was removed in favour of dom arguments.
The Substitution is initialized with its dom arguments.

<div data-montage-id="substitution">
  <div data-arg="one">One</div>
  <div data-arg="two">Two</div>
</div>

The data-args are the keys used to switch the to the specified element.

There is a programmatic API that can be used to add new switch elements to the repetition:

addSwitchElement(name, element)