Work-around for Repetition draw cycle
There appears to be a bug in the draw cycle where the Repetition
iterations have an opportunity to change between when canDraw() is
called and when draw() is called.  This goes from true to false if
the created and requested iteration counts are both zero (can draw) and
then an iteration is requested (can't draw until it is created).  This
request appears to be interposed between canDraw() and draw() presently.

This also make sure to retract iterations from their original positions
before injecting them at new positions.  This was fixed but must have
gotten lost in a shuffle.