Pass along key composer identifier with the event
The keyComposer system needs some work to dovetail better with the
activeTarget.

This change is the least invasive change to expose which keyComposer
identified the event. I am stopping short of any changes to eventManager
or the keyComposer itself as its warrants more consideration than I
can put in right now.