Swap Montage and application requires around in bootstrap
So that Montage is loaded with the application require, instead of
vice versa.

This fixes the issue where if an application and Montage depend on
the same package, npm would (correctly) install it in the
application's node_modules but, because the montage package was loaded
first, Montage would try and use a version in its node_modules.