Make Loader LOADED when main component enters document
It was possible for the Loader to be in a state where the main document was displayed in the screen (entered) but its stage was still PRELOADING.

Now we check the stage when the main component enters the document and update it if needed.

This solves the problem of seeing a logo and the main content on the top of each other because there was no way for the developer to create a css selector that would hide the logo after the main component was displayed.