Guard against unexpected keyPress action dispatch
The fairly generic handleKeyPress was not enough not ensure that the
keyPress was from the expected keyComposer.

I am opting to not use the identifier in this situation as it's null by
default meaning other keyComposers that did not have an identifier set
would continue to be indistinguishable. Checking the actual composer
avoids the issue.

The keyComposer system in general needs some review so I don't feel too
bad putting this in there now.