Change who claims the pointer based on interaction speed
Translate composers now have two additional properties:
- stealChildrenPointer
- stealChildrenPointerThreshold

When stealChildrenPointer is set to true on a translate composer, it will be able to claim the pointer even if a nested translated composer would have claim it otherwise.
This only happens when the initial reaction (the time between touchstart and touchend) is bellow a specific threshold (stealChildrenPointerThreshold).
stealChildrenPointerThreshold has a default value that suits both iPad and the Nexus 10.

This mechanism is meant to solve cases of nested translate composers like a vertical slider inside a vertical scroller.