Add missing properties to translateStart
When translateStart was fired by the FlowTranslateComposer it wasn't including translateX and translateY.
The MutableEvent that wraps the event fired creates a new constructor per event type. As an optimization it also stores that constructor the first time a specific type of event is fired.
If the flow is the first one to fire a translateStart event the shape of the MutableEvent constructor would be missing the translateX/translateY properties making future fires of translateStart by the TranslateComposer to miss those two properties.

This workaround just adds those two properties but a more effective way to solve this problem is to create a translateStart object that defines the needed interface.