Fix issue with capturing the pointer in touch
When two or more PressComposers were in play during a touchstart only the first one (the outer one) would capture the pointer. The others would just ignore it because the pointer was already captured.
This was preventing the inner most PressComposer to capture the pointer and handle it, which is the desired behavior.

This doesn't happen on mouse interactions because all PressComposers capture the pointer even if someone has already captured it before.(cherry picked from commit db482cb72202fae8b46dcd4748cf1b660a53e658)