Refactor Component._getDomArgument()
_getDomArgument was being incorrectly used by the substitution.
My guess is that this happened because of the generic name and functionality and lack of documentation.
It has now been renamed (as _getTemplateDomArgument) and repurposed to fit the specific need of the TemplateArgumentProvider.
A copy of the function was put into the Substitution (as _getSubstitutionDomArgument) with the necessary changes for it to work.

The code to find the data-arg is suspicious, I do not believe we need all that machinery, created a bug report about it to fix it in the near future.