Support component labels in defineBinding getPath
Previously, defineBinding and getPath did not support FRB `@label`
notation because they had no way to access peer components of the
object. @aadsm reveals that components have an `_ownerDocumentPart` with
that knowledge. This change bridges that gap. The upshot should be that
`sortPath` and `filterPath`, among other such expressions, should have
access to peer components now.