Avoid mouse event emulation bug
In Android browser, the emulated mouse events that come after the touch
do not always use the same target as the original touch events…
It looks like the target is recalculated at dispatch time, so if the
target of the touch event is moved out of the way by the time the mouse
event fires then it's target is wrong. 

This avoids the problem by only listening to one or the other on that
browser.