Don't stop drawing repetition if it started.
When the content of the repetition was changed in willDraw, it could get
into a state where it would abort in draw but its iterations that were
ready before the draw cycle drew anyway.
This fixes it by completion the draw while avoiding the newly created
iterations that are not yet ready.