visibility: collapse instead of display: none
Using visibility: collapse instead of display: none allows for the modal mask properties to be transitioned/animated using just CSS on show()/hide(). This can be useful if the user wants the modal mask to be animated when it is shown/hidden. 

More info about this property: https://developer.mozilla.org/en-US/docs/Web/CSS/visibility