Migrate from global require to mr
Fixes #1039

Many other names were considered including montageRequire (discarded
because it conflicts with a global variable used in Montage Testing),
applicationRequire (not specific enough to Montage), Montage.require
(because Montage is not a global and would be confused with and mask the
Montage constructor), and various others. `mr` balances specificity and
brevity.