Update to new TreeController
This new TreeController is backwards incompatible with the previous one.

The new controller enhances performance by only listening to changes in
the children of nodes that are both expanded and reachable. A node is
reachable if it is connected to the root by a path where all nodes
- root inclusive - are expanded.

Iterations have been removed and it is expected that views using the
controller will decide how to manage the iterations depending on their
rendering and culling methods.