Remove RangeController visibleIndexes/start/length
As a performance optimisation, now organizedContent is not enforced to
become a rangeContent, so if the content is not sorted, reversed or
filtered, the provided content will be the same object as
organizedContent.

This changes the behaviour of organizedContent, that before was ensured
to be an array, but now it might be null or undefined.

This affected abstract-select, that was accessing
organizedContent.length without checking for organizedContent existence
as it wasn’t needed before. It has been fixed by adding the checking.

It also affected a RangeController test and the test has been fixed too.