Fix Localizer not checking localStorage available
Localizer is using localStorage to store the preferred language. The
setItem function requires to be inside a try/catch in case the quota has
been exceeded to catch the quota exceeded error. This was causing
problems with Safari/iOS, that emits that error when you are browsing in
private mode as a way to avoid localStorage usage.