Adds FRB notion of reverter to mapping rules
In FRB syntax a binding can have either a converter or a reverter de-
fined.  When a reverter is defined the binding reverses the converter's
revert and convert function.

A mapping rule can now leverage this syntax by specifying a reverter
instead of a converter when defining the mapping rule.