Fix stride animation not happening case in Flow
This commit fixes a bug in flow-translate-composer that allows a flow
stride to stop in middle-of-stride points without starting a stride
animation. The bug was happening when flow scroll was between stride
points and a fast click / tap without pointer moves happened, then the stride
animation should kick in but it wasn't.