Make context available to ExpressionDataMapping rules
RawDataService.addRawData accepts an optional context
argument that can be used in the imperative
RawDataService.mapRawDataToObject. Previously, the
context was not accessible in ExpressionDataMapping rules so
mapRawDataToObject implementations that required it could not
be easily converted to ExpressionDataMapping. This commit
addresses that issue by assigning the context to scope.parameters
when the mapping rule is evaluated. Properties of the context can
be referenced in mapping rules with ‘$’ notation.